requirements.txt
Django===4.2.8
psycopg2-binary==2.9.9
django-redis==5.4.0
channels===3.0.5
SQLAlchemy===2.0.23
argon2-cffi===23.1.0
cryptography===42.0.8
django-cors-headers===4.3.1
pytz===2024.1 
requests===2.32.2
psycopg2===2.9.9